+++ tests/test_sklearn_pipeline.py
@@ -389,9 +389,7 @@ class TestSklearnPipeline(unittest.TestCase):
         model_onnx = convert_sklearn(
             model,
             "column transformer weights",
-            [("input", FloatTensorType([None, X.shape[1]]))],
-            dtype=numpy.float32,
-        )
+            [("input", FloatTensorType([None, X.shape[1]]))])
         self.assertIsNotNone(model_onnx)
         dump_data_and_model(
             X,
@@ -417,9 +415,7 @@ class TestSklearnPipeline(unittest.TestCase):
         model_onnx = convert_sklearn(
             model,
             "column transformer drop",
-            [("input", FloatTensorType([None, X.shape[1]]))],
-            dtype=numpy.float32,
-        )
+            [("input", FloatTensorType([None, X.shape[1]]))])
         self.assertIsNotNone(model_onnx)
         dump_data_and_model(
             X,
@@ -446,9 +442,7 @@ class TestSklearnPipeline(unittest.TestCase):
         model_onnx = convert_sklearn(
             model,
             "column transformer passthrough",
-            [("input", FloatTensorType([None, X.shape[1]]))],
-            dtype=numpy.float32,
-        )
+            [("input", FloatTensorType([None, X.shape[1]]))])
         self.assertIsNotNone(model_onnx)
         dump_data_and_model(
             X,
@@ -474,9 +468,7 @@ class TestSklearnPipeline(unittest.TestCase):
         model_onnx = convert_sklearn(
             model,
             "column transformer passthrough",
-            [("input", FloatTensorType([None, X.shape[1]]))],
-            dtype=numpy.float32,
-        )
+            [("input", FloatTensorType([None, X.shape[1]]))])
         self.assertIsNotNone(model_onnx)
         dump_data_and_model(
             X,
